First Functioning Version
This version of the app tracks your movement when you press the Start Run button, and stops tracking your movement when you press the End Run button. A polyline is drawn every 50 meters of movement, and the distance travelled is also listed when the End Run button is pressed for testing.