Added UIView for User Interaction
This commit includes a white UIView at the bottom of the MKMapView that allows the user to center their location on the map (not relative to the UIView quite yet), start a run, and stop a run. The functionality for ending a run and then starting a new one doesn't work (the MKPolyline from the previous run appears when a new one is started), but I'll fix that in a future commit.) I will also add a timer here soon.