Added Sharing and Workflow Enhancements
At this point, the app fulfills all the requirements of the assignment. The user can tap the share button to share a screenshot of the route they ran. This screenshot will include the new label that appears when the END RUN button is tapped; it shows how far the user ran in their previous run. Lastly, the START RUN button's title is changed to NEW RUN after the END RUN button is tapped to make the flow of the app make more sense.

I'm planning on making the code cleaner by adding some state management in a future commit.